--- backend/app/models/order.rb
@@ -1,11 +1,0 @@
-class Order < ApplicationRecord
-  # user は identity の所有。規約推論だと存在しない top-level User を探すため明示
-  belongs_to :user, class_name: "Identity::User"
-  has_many :order_items, dependent: :destroy
-  # coupon_use は promotion の所有。規約推論だと存在しない top-level CouponUse を探すため明示
-  has_one :coupon_use, class_name: "Promotion::CouponUse", dependent: :destroy
-
-  enum :status, { confirmed: "confirmed", cancelled: "cancelled" }
-
-  validates :order_number, presence: true, uniqueness: true
-end
