+++ backend/packs/catalog/app/models/catalog/product_variant.rb
@@ -3,8 +3,9 @@ module Catalog
     self.table_name = "product_variants"
 
     belongs_to :product
-    has_many :cart_items, dependent: :destroy
-    has_many :order_items, dependent: :nullify
+    # cart_items / order_items は ordering の所有。規約推論だと存在しない top-level を探すため明示
+    has_many :cart_items, class_name: "Ordering::CartItem", dependent: :destroy
+    has_many :order_items, class_name: "Ordering::OrderItem", dependent: :nullify
     has_one :stock, dependent: :destroy
 
     after_create :create_stock!
