+++ backend/packs/ordering/app/models/ordering/cart.rb
@@ -1,0 +1,9 @@
+module Ordering
+  class Cart < ApplicationRecord
+    self.table_name = "carts"
+
+    # user は identity の所有。規約推論だと存在しない top-level User を探すため明示
+    belongs_to :user, class_name: "Identity::User"
+    has_many :cart_items, dependent: :destroy
+  end
+end
