+++ backend/packs/ordering/app/models/ordering/order.rb
@@ -1,0 +1,15 @@
+module Ordering
+  class Order < ApplicationRecord
+    self.table_name = "orders"
+
+    # user は identity の所有。規約推論だと存在しない top-level User を探すため明示
+    belongs_to :user, class_name: "Identity::User"
+    has_many :order_items, dependent: :destroy
+    # coupon_use は promotion の所有。規約推論だと存在しない top-level CouponUse を探すため明示
+    has_one :coupon_use, class_name: "Promotion::CouponUse", dependent: :destroy
+
+    enum :status, { confirmed: "confirmed", cancelled: "cancelled" }
+
+    validates :order_number, presence: true, uniqueness: true
+  end
+end
