+++ backend/packs/promotion/app/models/promotion/coupon_use.rb
@@ -7,7 +7,8 @@ module Promotion
     belongs_to :coupon  # どのクーポンが使われたか
     # user は identity の所有。規約推論だと存在しない top-level User を探すため明示
     belongs_to :user, class_name: "Identity::User"  # 誰が使ったか（1ユーザー1回制限の判定に必要）
-    belongs_to :order   # どの注文で使われたか（キャンセル時にどの使用記録を削除するか特定するために必要）
+    # order は ordering の所有。規約推論だと存在しない top-level Order を探すため明示
+    belongs_to :order, class_name: "Ordering::Order"  # どの注文で使われたか（キャンセル時にどの使用記録を削除するか特定するために必要）
 
     enum :status, { unused: "unused", used: "used" }
